New IP Addresses
Added new IP addresses found VIA hurricane electric.

All addresses can be confirmed using their search function. Some of my
labels may not be apparent off the bat. In these cases, I checked the
whois, ARR, and DNS tabs to find the specific they are allocated to.
These are especially useful for when looking at the Shared Government
Services and Government Telecommunications and Informatics Services.

I also used the site ipinfo.io to confirm my additions. You can search
with ipinfo.io, but cannot put in ip ranges. Place IP in, and it will
display what network it belongs to, and the IP’s ASN, allowing you to
find more information regarding the ranges.

As for the JSON format, you are using two separate types of formats
(Government of Canada to Environment Canada is one format, Library and
Archives Canada to Federal Court of Canada is another). If you want to
tweet @AniMud with which format you prefer, I can go ahead and make the
changes to the file, along with my current and future additions.

Not much was added today as I’m running out the door. But hopefully
this helps.